social-app/server/requirements.txt
asgiref==3.11.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
bleach==6.3.0
cbor2==5.8.0
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.6
cloudinary==1.44.1
constantly==23.10.4
cryptography==46.0.5
daphne==4.2.1
dj-database-url==3.1.2
Django==4.2.29
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
msgpack==1.1.2
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
pyOpenSSL==26.0.0
python-dotenv==1.2.2
requests==2.32.5
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.12.0
urllib3==2.6.3
webencodings==0.5.1
zope.interface==8.2
whitenoise==6.8.2
django-ratelimit==4.1.0
channels-redis==4.2.0
redis==5.0.1
